Remove file db (#237)
* Revert "try(backend): revert to global state"

This reverts commit 78385c5f9468dc151d356bc1f23f6c9c913fce2a.

* build(backend): bump version

* feat(backend): make scheduler database memory mode

* feat(backend): make reminder naive date

* feat(frontend): use date input for setting reminder

* feat(backend): send media reminders

+++ apps/backend/src/config.rs
@@ -360,12 +360,12 @@ impl IsFeatureEnabled for FileStorageConfig {
 #[config(rename_all = "snake_case", env_prefix = "SCHEDULER_")]
 pub struct SchedulerConfig {
     /// The url to the SQLite database where job related data needs to be stored.
-    #[setting(default = format!("sqlite:/data/{}-scheduler.db?mode=rwc", PROJECT_NAME))]
+    #[setting(default = "sqlite::memory:")]
     pub database_url: String,
     /// The number of jobs to process every 5 seconds when updating metadata in
     /// the background.
     #[setting(default = 5)]
-    pub rate_limit_num: i32,
+    pub rate_limit_num: u64,
     /// Deploy a job every x hours that performs user cleanup and summary
     /// calculation.
     #[setting(default = 12)]
